+++ cmd/src/main.go
@@ -9,6 +9,8 @@ import (
 	"os/user"
 	"path/filepath"
 	"strings"
+
+	"github.com/pkg/errors"
 )
 
 const usageText = `src is a tool that provides access to Sourcegraph instances.
@@ -24,7 +26,6 @@ Environment variables
 
 The options are:
 
-	-endpoint=                       specifies the endpoint to use e.g. "https://sourcegraph.com" (overrides SRC_ENDPOINT if set)
 	-v                               print verbose output
 
 The commands are:
@@ -47,9 +48,11 @@ Use "src [command] -h" for more information about a command.
 `
 
 var (
+	verbose = flag.Bool("v", false, "print verbose output")
+
+	// The following arguments are deprecated which is why they are no longer documented
 	configPath = flag.String("config", "", "")
 	endpoint   = flag.String("endpoint", "", "")
-	verbose    = flag.Bool("v", false, "print verbose output")
 )
 
 // commands contains all registered subcommands.
@@ -76,14 +79,14 @@ func readConfig() (*config, error) {
 	cfgPath := *configPath
 	userSpecified := *configPath != ""
 
-	user, err := user.Current()
+	u, err := user.Current()
 	if err != nil {
 		return nil, err
 	}
 	if !userSpecified {
-		cfgPath = filepath.Join(user.HomeDir, "src-config.json")
+		cfgPath = filepath.Join(u.HomeDir, "src-config.json")
 	} else if strings.HasPrefix(cfgPath, "~/") {
-		cfgPath = filepath.Join(user.HomeDir, cfgPath[2:])
+		cfgPath = filepath.Join(u.HomeDir, cfgPath[2:])
 	}
 	data, err := ioutil.ReadFile(os.ExpandEnv(cfgPath))
 	if err != nil && (!os.IsNotExist(err) || userSpecified) {
@@ -96,23 +99,39 @@ func readConfig() (*config, error) {
 		}
 	}
 
+	envToken := os.Getenv("SRC_ACCESS_TOKEN")
+	envEndpoint := os.Getenv("SRC_ENDPOINT")
+
+	if userSpecified {
+		// If a config file is present, either zero or both environment variables must be present.
+		// We don't want to partially apply environment variables.
+		if envToken == "" && envEndpoint != "" {
+			return nil, errConfigMerge
+		}
+		if envToken != "" && envEndpoint == "" {
+			return nil, errConfigMerge
+		}
+	}
+
 	// Apply config overrides.
-	if envToken := os.Getenv("SRC_ACCESS_TOKEN"); envToken != "" {
+	if envToken != "" {
 		cfg.AccessToken = envToken
 	}
-	if *endpoint != "" {
-		cfg.Endpoint = *endpoint
-	}
-	if cfg.Endpoint == "" {
-		if endpoint := os.Getenv("SRC_ENDPOINT"); endpoint != "" {
-			cfg.Endpoint = endpoint
-		}
+	if envEndpoint != "" {
+		cfg.Endpoint = envEndpoint
 	}
 	if cfg.Endpoint == "" {
 		cfg.Endpoint = "https://sourcegraph.com"
+	}
+
+	// Lastly, apply endpoint flag if set
+	if endpoint != nil && *endpoint != "" {
+		cfg.Endpoint = *endpoint
 	}
 
 	cfg.Endpoint = strings.TrimSuffix(cfg.Endpoint, "/")
 
 	return &cfg, nil
 }
+
+var errConfigMerge = errors.New("when using a configuration file, zero or all environment variables must be set")